requirements.txt
requests
pandas
numpy
matplotlib
jieba
pdfplumber
python-docx
nltk
sentence-transformers
scikit-learn
torch
transformers
langchain
langchain-openai
langchain-huggingface
ragas
evaluate
datasets
